+++ mail/mail-uniffi/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
  - [ET-548]: Report an issue feature is fully implemeneted
+ - SystemLabel has two new locations: `Blocked`, `Pinned`.
 
 
 ## [0.62.1] - 2025-03-24